requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-Session==0.8.0
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyserial==3.5
requests==2.31.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
